refactor: modularize I-MPPI simulation
- Moved FSMI logic to `examples/i_mppi_modules/fsmi.py`
- Moved Biased MPPI planner to `examples/i_mppi_modules/planner.py`
- Moved environment definition to `examples/i_mppi_modules/environment.py`
- Updated `examples/i_mppi_simulation.py` to use the new modules.
